# backend/openapi.json
{
  "openapi": "3.0.0",
  "paths": {
...
    "/api/admin/providers": {
      "get": {
        "operationId": "AiController_providers",
        "parameters": [],
        "responses": {
          "200": {
            "description": ""
          }
        },
        "security": [
          {
            "bearer": []
          }
        ],
        "tags": [
          "console · intelligence artificielle"
        ]
      },
      "post": {
        "operationId": "AiController_create",
        "parameters": [],
        "responses": {
          "201": {
            "description": ""
          }
        },
        "security": [
          {
            "bearer": []
          }
        ],
        "tags": [
          "console · intelligence artificielle"
        ]
      }
    },
    "/api/admin/providers/{id}/key": {
      "put": {
        "operationId": "AiController_rotate",
        "parameters": [
          {
            "name": "id",
            "required": true,
            "in": "path",
            "schema": {
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "description": ""
          }
        },
        "security": [
          {
            "bearer": []
          }
        ],
        "tags": [
          "console · intelligence artificielle"
        ]
      }
    },
    "/api/admin/providers/{id}/test": {
      "post": {
        "operationId": "AiController_testOne",
        "parameters": [
          {
            "name": "id",
            "required": true,
            "in": "path",
            "schema": {
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "description": ""
          }
        },
        "security": [
          {
            "bearer": []
          }
        ],
        "tags": [
          "console · intelligence artificielle"
        ]
      }
    },
    "/api/admin/providers/test-all": {
      "post": {
        "operationId": "AiController_testAll",
        "parameters": [],
        "responses": {
          "200": {
            "description": ""
          }
        },
        "security": [
          {
            "bearer": []
          }
        ],
        "tags": [
          "console · intelligence artificielle"
        ]
      }
    },
    "/api/admin/models": {
      "get": {
        "operationId": "AiController_models",
        "parameters": [
          {
            "name": "provider",
            "required": true,
            "in": "query",
            "schema": {
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "description": ""
          }
        },
        "security": [
          {
            "bearer": []
          }
        ],
        "tags": [
          "console · intelligence artificielle"
        ]
      }
    },
    "/api/admin/models/{id}": {
      "patch": {
        "operationId": "AiController_patchModel",
        "parameters": [
          {
            "name": "id",
            "required": true,
            "in": "path",
            "schema": {
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "description": ""
          }
        },
        "security": [
          {
            "bearer": []
          }
        ],
        "tags": [
          "console · intelligence artificielle"
        ]
      }
    },
    "/api/admin/assignments": {
      "get": {
        "operationId": "AiController_assignments",
        "parameters": [],
        "responses": {
          "200": {
            "description": ""
          }
        },
        "security": [
          {
            "bearer": []
          }
        ],
        "tags": [
          "console · intelligence artificielle"
        ]
      },
      "put": {
        "operationId": "AiController_putAssignments",
        "parameters": [],
        "responses": {
          "200": {
            "description": ""
          }
        },
        "security": [
          {
            "bearer": []
          }
        ],
        "tags": [
          "console · intelligence artificielle"
        ]
      }
    },
    "/api/admin/usage/month": {
      "get": {
        "operationId": "AiController_month",
        "parameters": [],
        "responses": {
          "200": {
            "description": ""
          }
        },
        "security": [
          {
            "bearer": []
          }
        ],
        "tags": [
          "console · intelligence artificielle"
        ]
      }
    },
    "/api/admin/usage": {
      "get": {
        "operationId": "AiController_series",
        "parameters": [
          {
            "name": "from",
            "required": true,
            "in": "query",
            "schema": {
              "type": "string"
            }
          },
          {
            "name": "to",
            "required": true,
            "in": "query",
            "schema": {
              "type": "string"
            }
          },
          {
            "name": "projectId",
            "required": true,
            "in": "query",
            "schema": {
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "description": ""
          }
        },
        "security": [
          {
            "bearer": []
          }
        ],
        "tags": [
          "console · intelligence artificielle"
        ]
      }
    },
    "/api/admin/usage/export.csv": {
      "get": {
        "operationId": "AiController_exportCsv",
        "parameters": [
          {
            "name": "from",
            "required": true,
            "in": "query",
            "schema": {
              "type": "string"
            }
          },
          {
            "name": "to",
            "required": true,
            "in": "query",
            "schema": {
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "description": ""
          }
        },
        "security": [
          {
            "bearer": []
          }
        ],
        "tags": [
          "console · intelligence artificielle"
        ]
      }
    },
    "/api/admin/budget-thresholds": {
      "get": {
        "operationId": "AiController_thresholds",
        "parameters": [],
        "responses": {
          "200": {
            "description": ""
          }
        },
        "security": [
          {
            "bearer": []
          }
        ],
        "tags": [
          "console · intelligence artificielle"
        ]
      }
    },
    "/api/admin/budget-thresholds/{id}": {
      "put": {
        "operationId": "AiController_putThreshold",
        "parameters": [
          {
            "name": "id",
            "required": true,
            "in": "path",
            "schema": {
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "description": ""
          }
        },
        "security": [
          {
            "bearer": []
          }
        ],
        "tags": [
          "console · intelligence artificielle"
        ]
      }
    },
...
  },
  "info": {
    "title": "RISE — API Cockpit et Console Admin",
    "description": "API du Cockpit (`/api`, `/api/projects/{projectId}/…`) et de la Console Admin (`/api/admin/…`). Erreurs : `{ code, message, fields?, usages? }`. Verrouillage optimiste : `If-Match` / `ETag`.",
    "version": "1.0.0",
    "contact": {}
  },
...
  "servers": [],
  "components": {
    "securitySchemes": {
      "bearer": {
        "scheme": "bearer",
        "bearerFormat": "JWT",
        "type": "http"
      }
    },
...
  }
}